Extract the core tool into lib.rs
